TSDK-279 CredentiallerInterpreter (#14)
* First pass at refactoring Credentialler. 

Notable differences 1) validation will do syntax and authorization validation. 2) Signatures updated to be simpler since at this moment in time its not clear why we would need errors during the prove stage. It was originally added because of a Syntax check that no longer upholds

* Removed hardcoded usage of the MockDataApi

* Cleaned up the private functions within CredentiallerInterpreter; updated documentation and also replaced Option[Proof] with Proof() (i.e, Proof.Value.Empty)

* Removed hardcoded Id from helper functions. Instead use F context

* Ran formatter

* Moved MockDataApi to the tests directory. Also cleaned up MockHelpers

MockDataApi will work in conjunction with MockHelpers. This is in preparation for CredentiallerInterpreterSpec. Ensured all existing tests pass

* Implemented basic unit tests for credentialler

* Ran formatter