TSDK-352 Added ExtendedEd25519 to Ecosystem Crypto (#29)
* First pass at preparing existing code to accept ExtendedEd25519 

- replaced usage of ByteVector with models and ByteArrays
- minor updates to implementation (such as extracted a method into parts so it can be easily overriden)

* Added ExtendedEd25519 Scheme

* Ran formatter

* refactored yesterday's work to not use the compiler's Sized but added
require assertions for the types constructors

* Added comments and ran formatter

* Fixed existing tests so that they all pass 

Do to the refactor of Ed25519 and removal of ByteVector, some tests failed.

* Added tests for ExtendedEd25519

Minor update to Ed25519 tests as well to utilizes "equals" method for the Keys. 3 new tests are still failing

* The 3 failing tests required KeyInitializer to be implemented for ExtendedEd25519

Thus partial work of TSDK-354 was done. All existing tests pass + the new Signing tests for ExtendedEd25519

* Added tests for KeyInitializer

Because ExtendedEd25519 was added to KeyInitializer

* Minor changes

* Updated Quivr and Protobuf depedencies